Extract openapi codegen to plugin
