+++ build-logic/build.gradle.kts
@@ -6,4 +6,5 @@ plugins {
 
 dependencies {
     implementation("org.jetbrains.kotlin:kotlin-gradle-plugin:1.9.23")
+    implementation("org.openapitools:openapi-generator-gradle-plugin:7.4.0")
 }
