+++ build-logic/src/main/kotlin/com/gabrielfeo/kotlin-jvm-library.gradle.kts
@@ -2,4 +2,5 @@ package com.gabrielfeo
 
 plugins {
     id("org.jetbrains.kotlin.jvm")
+    `java-library`
 }
